third_party/WebKit/LayoutTests/accessibility/input-type-password-selection.html

<!DOCTYPE HTML PUBLIC>
<html>
<head>
<script src="../resources/js-test.js"></script>
</head>
<body>

<p>This tests that the password fields return their selection.</p>

<input id="password" type="password" value="secret">

<div id="console"></div>
<script>
    if (window.testRunner && window.accessibilityController) {
        var password = document.getElementById("password");
        var axPassword = accessibilityController.accessibleElementById("password");

        password.setSelectionRange(0, 0);
        shouldBe("axPassword.selectedTextRange", "\"{0, 0}\"");

        password.setSelectionRange(6, 6);
        shouldBe("axPassword.selectedTextRange", "\"{6, 0}\"");

        password.setSelectionRange(2, 6);
        shouldBe("axPassword.selectedTextRange", "\"{2, 4}\"");
    }

</script>

</body>
</html>
