third_party/WebKit/LayoutTests/accessibility/input-type-password-selection-expected.txt

This tests that the password fields return their selection.


PASS axPassword.selectedTextRange is "{0, 0}"
PASS axPassword.selectedTextRange is "{6, 0}"
PASS axPassword.selectedTextRange is "{2, 4}"
PASS successfullyParsed is true

TEST COMPLETE
